move data and fix path
