+++ packages/hono-takibi/data/pet-store-openapi.ts
@@ -1,4 +1,4 @@
-import type { OpenAPISpec } from '../type'
+import type { OpenAPISpec } from '../src/type'
 
 export const petStoreOpenAPI: OpenAPISpec = {
   openapi: '3.0.3',
