+++ packages/hono-takibi/data/test-data.ts
@@ -1,4 +1,4 @@
-import type { Config } from '../config'
+import type { Config } from '../src/config'
 
 export const DEFAULT_CONFIG: Config = {
   schema: {
