+++ packages/hono-takibi/src/generator/zod-openapi-hono/app/index.test.ts
@@ -1,9 +1,9 @@
-import type { OpenAPISpec } from '../../../type'
-import type { Config } from '../../../config'
 import { describe, expect, it } from 'vitest'
 import { generateApp } from './'
+import { honoRestOpenAPI } from '../../../../data/hono-rest-openapi'
 import { DEFAULT_CONFIG } from '../../../config'
-import { honoRestOpenAPI } from '../../../data/hono-rest-openapi'
+import type { OpenAPISpec } from '../../../type'
+import type { Config } from '../../../config'
 
 const generateAppTestCases: {
   openAPISpec: OpenAPISpec
