+++ packages/hono-takibi/src/generator/zod-openapi-hono/handler/import/generate-import-handlers.test.ts
@@ -1,7 +1,7 @@
-import type { Config } from '../../../../config'
-import { DEFAULT_CONFIG } from '../../../../config'
 import { describe, expect, it } from 'vitest'
 import { generateImportHandlers } from './generate-import-handlers'
+import { DEFAULT_CONFIG } from '../../../../config'
+import type { Config } from '../../../../config'
 
 const generateImportHandlersTestCases: {
   handlerImportsMap: { [fileName: string]: string[] }
