+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/component/allof/process/process-alllof.test.ts
@@ -1,8 +1,8 @@
-import type { Schema } from '../../../../../../type'
-import type { Config } from '../../../../../../config'
-import { DEFAULT_CONFIG } from '../../../../../../config'
 import { describe, expect, it } from 'vitest'
 import { processAllOf } from './process-alllof'
+import { DEFAULT_CONFIG } from '../../../../../../config'
+import type { Schema } from '../../../../../../type'
+import type { Config } from '../../../../../../config'
 
 const processAllOfTestCases: {
   allOf: Schema[]
