+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/component/schema/generate-schemas-export.test.ts
@@ -1,7 +1,7 @@
-import type { Config } from '../../../../../config'
 import { describe, expect, it } from 'vitest'
 import { generateSchemasExport } from './generate-schemas-export'
-import { DEFAULT_CONFIG, PASCAL_CASE_CONFIG } from '../../../../../data/test-data'
+import { DEFAULT_CONFIG, PASCAL_CASE_CONFIG } from '../../../../../../data/test-data'
+import type { Config } from '../../../../../config'
 
 const generateSchemasExportTestCases: {
   orderedSchemas: string[]
