+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/generate-params-object.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from 'vitest'
 import { generateParamsObject } from './generate-params-object'
 import type { ParamsObject, Parameters } from '../../../../../type'
-import { DEFAULT_CONFIG, type Config } from '../../../../../config'
+import type { Config } from '../../../../../config'
+import { DEFAULT_CONFIG } from '../../../../../config'
 
 const generateParamsObjectTestCases: {
   parameters: Parameters[]
