+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/generate-request-parameter.test.ts
@@ -1,8 +1,8 @@
-import type { Parameters, RequestBody } from '../../../../../type'
-import type { Config } from '../../../../../config'
-import { DEFAULT_CONFIG } from '../../../../../data/test-data'
 import { describe, expect, it } from 'vitest'
 import { generateRequestParameter } from './generate-request-parameter'
+import { DEFAULT_CONFIG } from '../../../../../../data/test-data'
+import type { Parameters, RequestBody } from '../../../../../type'
+import type { Config } from '../../../../../config'
 
 const generateRequestParameterTestCases: {
   parameters: Parameters[] | undefined
