+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/response/generate-response-schema.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from 'vitest'
 import { generateResponseSchema } from './generate-response-schema'
+import { DEFAULT_CONFIG } from '../../../../../config'
 import type { Responses } from '../../../../../type'
-import { DEFAULT_CONFIG, type Config } from '../../../../../config'
+import type { Config } from '../../../../../config'
 
 const generateResponseSchemaTestCases: {
   responses: Responses
