+++ packages/hono-takibi/src/generator/zod/generate-zod-record.test.ts
@@ -1,8 +1,8 @@
-import type { Format, Type } from '../../type'
-import type { Config } from '../../config'
-import { DEFAULT_CONFIG } from '../../data/test-data'
 import { describe, expect, it } from 'vitest'
 import { generateZodRecord } from './generate-zod-record'
+import { DEFAULT_CONFIG } from '../../../data/test-data'
+import type { Format, Type } from '../../type'
+import type { Config } from '../../config'
 
 const generateZodRecordTestCases: {
   additionalProperties: { type: Type; format: Format }
