+++ packages/hono-takibi/src/generator/zod/property/generate-zod-properties-schema.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
-import type { Schema } from '../../..type'
-import type { Config } from '../../../config'
 import { generateZodPropertiesSchema } from './generate-zod-properties-schema'
-import { DEFAULT_CONFIG } from '../../../data/test-data'
+import { DEFAULT_CONFIG } from '../../../../data/test-data'
+import type { Schema } from '../../../type'
+import type { Config } from '../../../config'
 
 const generateZodPropertiesSchemaTestCases: {
   properties: Record<string, Schema>
