+++ packages/hono-takibi/src/generator/zod/property/generate-zod-property-schema.test.ts
@@ -1,8 +1,9 @@
 import { describe, expect, it } from 'vitest'
 import { generatePropertySchema } from './generate-zod-property-schema'
-import type { Schema } from '../../..type'
-import { DEFAULT_CONFIG, type Config } from '../../../config'
-import { PASCAL_CASE_CONFIG } from '../../../data/test-data'
+import { DEFAULT_CONFIG } from '../../../config'
+import { PASCAL_CASE_CONFIG } from '../../../../data/test-data'
+import type { Schema } from '../../../type'
+import type { Config} from '../../../config'
 
 const generatePropertySchemaTestCases: {
   schema: Schema
