+++ packages/hono-takibi/src/generator/zod/sub/generate-zod-schema-from-sub-schema.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 import { generateZodSchemaFromSubSchema } from './generate-zod-schema-from-sub-schema'
 import { DEFAULT_CONFIG } from '../../../config'
-import type { Schema } from '../../..type'
+import type { Schema } from '../../../type'
 import type { Config } from '../../../config'
 const generateZodSchemaFromSubSchemaTestCases: {
   subSchema: Schema
